+++ gui_agent/src/main.js
@@ -68,6 +68,7 @@ import "tabulator-tables/dist/css/tabulator.min.css";
 import "tabulator-tables/dist/css/tabulator_bulma.css";
 import "@yuneta/gobj-ui/src/tabulator.css";   // global Tabulator theme fixes
 import "./app.css";
+import "./login.css";
 
 import app_config from "./app_config.json";